Refactor: move PlotDistribution out of distribution experiment
It is a more gerenally useful function that can now be used in other
experiments.
